decoder: add hasKey tag
The `hasKey` fixes a problem where a struct with a list field where the
list elements have a field with the `key` tag would receive the wrong
key value if the block was defined without a label.

For example, an struct like this:

```
type A struct {
  Bs []*B `hcl:"b"`
}

type B struct {
  C string `hcl:",key"`
  D string
}
```

And an input like this:

```
b {
  d = "e"
}

b "z" {
  d = "y"
}
```

Generates an incorrect output:

```
&A {
  Bs: []*B{
    {
      C: "d", // wut?
      D: "e",
    },
    {
      C: "z", // ok
      D: "y",
    },
  },
}
```

This happens because when the decoder is traversing the AST it doesn't
know if a nested field expects a key value, and so it makes a wrong
assumption about what the key value should be.

Internally, each item in the AST has keys and another node of the tree.
For blocks, the keys are the block type plus its labels. In the example
above, the first block has one key, the identifier b. The second block
as two keys: the identifier b and a string label "z".

When the file is processed, the decoder receives all the blocks in the
root and process then based on each of fields in A.

Since the first field is a list of Bs, it filters from the root the
blocks that have a `b` identifier as the first key. This filtering
process also removes the `b` entry from the item's keys.

This leaves blocks without a label with no keys left, and considered to
be an exact match. Blocks that do have labels keep them as leftover
keys, and are considered prefix matches.

Exact matches have their children processed, not the item itself. This
creates the problem described here because the children will have an
unrelated value as key. In the example above, the block without label
receives `d` as key because that's the first key of the sub-element.

The `hasKey` tag signals the decoder that the list elements expect to
receive a key, even an empty one. And so this commit turns exact matches
into prefix matches, forcing the item itself to be processed instead of
its children.

Adding this tag to the example generates the expected output:

```
type A struct {
  Bs []*B `hcl:"b,hasKey"`
}

type B struct {
  C string `hcl:",key"`
  D string
}
```

```
&A {
  Bs: []*B{
    {
      C: "", // fixed
      D: "e",
    },
    {
      C: "z", // ok
      D: "y",
    },
  },
}
```